+++ src/sheets/core.ts
@@ -39,6 +39,46 @@ sheet.glyphs.set("cannon-town", (canvas: SVGContainer) => {
     return group;
 });
 
+sheet.glyphs.set("house", (canvas: SVGContainer) => {
+    const group = canvas.group()
+        .id("house")
+        .attr("data-cellsize", 80);
+    group.rect(80, 80).fill({opacity: 0});
+    const house = group.group();
+    house.rect(66.07, 21.700001).x(0.70999998).y(17.610001)
+        .stroke({width: 1, color: "black"})
+        .fill("none")
+        .attr("data-playerfill", true);
+    house.polygon("61.74,0.5 6.78,0.5 0.71,17.61 66.78,17.61")
+        .stroke({width: 1, color: "black"})
+        .fill("none")
+        .attr("data-playerfill", true);
+    house.translate(6.2746953, 40.189999);
+    return group;
+});
+
+sheet.glyphs.set("palace", (canvas: SVGContainer) => {
+    const group = canvas.group()
+        .id("palace")
+        .attr("data-cellsize", 80);
+    group.rect(80, 80).fill({opacity: 0});
+    const palace = group.group();
+    palace.polygon("64.94,33.45 33.35,13.89 33.35,0.93 61.38,19.52")
+        .stroke({width: 1, color: "black"})
+        .fill("none")
+        .attr("data-playerfill", true);
+    palace.polygon("0.86,33.45 33.35,13.89 33.35,0.93 5.38,18.86")
+        .stroke({width: 1, color: "black"})
+        .fill("none")
+        .attr("data-playerfill", true);
+    palace.polygon("64.94,57.38 64.94,33.45 33.35,13.89 0.86,33.45 0.86,57.38")
+        .stroke({width: 1, color: "black"})
+        .fill("none")
+        .attr("data-playerfill", true);
+    palace.translate(7.1370414, 22.12);
+    return group;
+});
+
 sheet.glyphs.set("piece", (canvas: SVGContainer) => {
     const group = canvas.group()
         .id("piece")
@@ -102,6 +142,28 @@ sheet.glyphs.set("piece-square", (canvas: SVGContainer) => {
         .fill("#fff")
         .stroke({width: 2, color: "#000"})
         .center(sheet.cellsize / 2, sheet.cellsize / 2);
+    return group;
+});
+
+sheet.glyphs.set("tower", (canvas: SVGContainer) => {
+    const group = canvas.group()
+        .id("tower")
+        .attr("data-cellsize", 80);
+    group.rect(80, 80).fill({opacity: 0});
+    const tower = group.group();
+    tower.polygon("32.5,0.9 60.05,18.12 64.61,32.12 32.5,12.23")
+        .stroke({width: 1, color: "black"})
+        .fill("none")
+        .attr("data-playerfill", true);
+    tower.polygon("32.5,0.9 4.94,17.46 0.83,32.12 32.5,12.23")
+        .stroke({width: 1, color: "black"})
+        .fill("none")
+        .attr("data-playerfill", true);
+    tower.polygon("64.61,32.12 32.5,12.23 0.83,32.12 0.83,78.46 64.61,78.46")
+        .stroke({width: 1, color: "black"})
+        .fill("none")
+        .attr("data-playerfill", true);
+    tower.translate(7.2461984, 1.04);
     return group;
 });